PyScope/PyModule cleanup
removed PyScopeManager
merged PyScope into PyModule
minor behavioral changes
